Create new www.dashboard_networks folder. Refactor
