+++ lemarche/templates/dashboard/home_buyer.html
@@ -124,7 +124,7 @@
                             </div>
                         </div>
                         <div class="card-footer pt-0 bg-white text-right">
-                            <a href="{% url 'dashboard:profile_network_detail' user.partner_network.slug %}" id="dashboard_network_detail" class="btn btn-link btn-ico">
+                            <a href="{% url 'dashboard_networks:detail' user.partner_network.slug %}" id="dashboard_network_detail" class="btn btn-link btn-ico">
                                 <span>Animer mon réseau</span>
                                 <i class="ri-arrow-right-s-line ri-xl"></i>
                             </a>
