+++ lemarche/templates/networks/dashboard_network_detail.html
@@ -48,7 +48,7 @@
                         </p>
                     </div>
                     <div class="card-footer pt-0 bg-white text-right">
-                        <a href="{% url 'dashboard:profile_network_siae_list' network.slug %}" id="dashboard-network-siae-list-btn" class="btn btn-link btn-ico">
+                        <a href="{% url 'dashboard_networks:siae_list' network.slug %}" id="dashboard-network-siae-list-btn" class="btn btn-link btn-ico">
                             <span>Voir la liste</span>
                             <i class="ri-arrow-right-s-line ri-xl"></i>
                         </a>
@@ -67,7 +67,7 @@
                         </p>
                     </div>
                     <div class="card-footer pt-0 bg-white text-right">
-                        <a href="{% url 'dashboard:profile_network_tender_list' network.slug %}" id="dashboard-network-tender-list-btn" class="btn btn-link btn-ico">
+                        <a href="{% url 'dashboard_networks:tender_list' network.slug %}" id="dashboard-network-tender-list-btn" class="btn btn-link btn-ico">
                             <span>Voir les opportunités</span>
                             <i class="ri-arrow-right-s-line ri-xl"></i>
                         </a>
