+++ lemarche/templates/tenders/_list_item_network.html
@@ -1,6 +1,6 @@
 {% load static humanize %}
 
-<div class="card c-card c-card--marche c-card--link siae-card" role="button" data-url="{% url 'dashboard:profile_network_tender_detail' network.slug tender.slug %}">
+<div class="card c-card c-card--marche c-card--link siae-card" role="button" data-url="{% url 'dashboard_networks:tender_detail' network.slug tender.slug %}">
     <div class="card-body">
         <div class="row">
             <div class="col-md-8" style="border-right:1px solid;">
@@ -51,7 +51,7 @@
                             <i class="ri-focus-2-line font-weight-light"></i>&nbsp;{{ tender.network_siae_email_send_count|default:0 }} adhérent{{ tender.network_siae_email_send_count|pluralize }} notifié{{ tender.network_siae_email_send_count|pluralize }}
                         </p>
                         {% if tender.network_siae_email_send_count %}
-                            <a href="{% url 'dashboard:profile_network_tender_siae_list' network.slug tender.slug %}" id="dashboard-network-tender-show-siae-list-btn" class="btn btn-sm btn-primary">
+                            <a href="{% url 'dashboard_networks:tender_siae_list' network.slug tender.slug %}" id="dashboard-network-tender-show-siae-list-btn" class="btn btn-sm btn-primary">
                                 Voir la liste
                             </a>
                         {% endif %}
